Concatenate 'N' variables

Concatenate 'N' variables using the char pipe: <|>. This function is used when there is the need of measuring the mutual information and/or the information gain between 'N' input variables an against a target variable. This function makes sense when it is used based on categorical data.

concatenate_n_vars(data, vars)

Arguments

data
data frame containing the two variables to concatenate

vars
character vector containing all variables to concatenate

Value

vector containing the concatenated values for the given variables

Examples

if (FALSE) {
new_variable=concatenate_n_vars(mtcars, c("cyl", "disp"))
# Checking new variable
head(new_variable)
}